requirements.txt
crewai>=0.11.0
python-dotenv>=1.0.0
jira>=3.5.1
langchain>=0.1.0
langchain-openai>=0.0.2